Gallery mode: Ignore hidden source elements.
This makes the gallery mode compatible with all sorts of client side filtering. As a side effect, the list of source elements used for gallery mode is now determined at the moment the lightbox is openend (in earlier versions it was determined when lighter box was initialized).

This may have a slight performance impact, but it should be negligible.